refactor: Rename legacy statedb package
