+++ x/evm/legacystatedb/access_list.go
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with the go-ethereum library. If not, see <http://www.gnu.org/licenses/>.
 
-package statedb_legacy
+package legacystatedb
 
 import (
 	"github.com/ethereum/go-ethereum/common"
