+++ x/evm/legacystatedb/state_object.go
@@ -13,7 +13,7 @@
 //
 // You should have received a copy of the GNU Lesser General Public License
 // along with the Ethermint library. If not, see https://github.com/evmos/ethermint/blob/main/LICENSE
-package statedb_legacy
+package legacystatedb
 
 import (
 	"bytes"
